+++ Slotmachine/src/main/java/com/example/casino/factory/ISlotmachineGameFactory.java
@@ -4,7 +4,9 @@ import com.example.casino.model.ISlotmachineGameEntity;
 import com.example.casino.utility.ErrorWrapper;
 import com.example.casino.utility.Result;
 
+import java.math.BigDecimal;
+
 public interface ISlotmachineGameFactory {
 
-    Result<ISlotmachineGameEntity, ErrorWrapper> createSlotmachine(long userId, double betAmount, double winAmount, boolean isWinning, String slotResult);
+    Result<ISlotmachineGameEntity, ErrorWrapper> createSlotmachine(long userId, BigDecimal betAmount, BigDecimal winAmount, boolean isWinning, String slotResult);
 }
