+++ Slotmachine/src/main/java/com/example/casino/factory/SlotmachineGameFactory.java
@@ -6,13 +6,14 @@ import com.example.casino.utility.ErrorWrapper;
 import com.example.casino.utility.Result;
 import org.springframework.stereotype.Component;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Component
 public class SlotmachineGameFactory implements ISlotmachineGameFactory {
 
     @Override
-    public Result<ISlotmachineGameEntity, ErrorWrapper> createSlotmachine(long userId, double betAmount, double winAmount, boolean isWinning, String slotResult){
+    public Result<ISlotmachineGameEntity, ErrorWrapper> createSlotmachine(long userId, BigDecimal betAmount, BigDecimal winAmount, boolean isWinning, String slotResult){
         LocalDateTime timestamp = LocalDateTime.now();
         return SlotmachineGameEntity.create(userId, betAmount, winAmount, isWinning, slotResult, timestamp);
     }
